Client: Don't require camera or mic.
Previously it was assumed that a camera and microphone were available,
and the client failed silently if either a camera or mic were not
available. Now neither a camera nor mic should be necessary to receive
media from peers or use screen sharing. Also, various bugs with
synchronizing button states have been fixed.