Implement dynamic layout for video stage.
The video stage can now display multiple video feeds within a fixed
area. The feeds are layed out to maximize the total area covered within
the video stage. They layout is fully responsive and recomputed any time
the screen is resized or a video feed is added or removed.

The video stage layout uses Flexbox (the video stage has `display: flex`
with `flex-wrap: wrap`). Layout works by first computing how many rows
and columns are needed to include all video feeds. Once the minimum
number of rows and columns is known, a scaling factor is then computed
and applied to calculate the height and width of each video feed. With
the proper scaling factor, the video feeds can simply be added to the
Flexbox and are layed out in the appropriate number of (implicit) rows
and columns.