Fix VideoFeed scaling on Sidebar toggle.
VideoFeeds only rescale on window resize events or React re-renderings,
so toggling the sidebar would not previously cause VideoFeeds to
rescale. This implementation is not ideal, but works. In the future, it
may be possible to use a Resize Observer to listen for changes in the
dimensions of the VideoStage (see
https://drafts.csswg.org/resize-observer-1/).