Stop preview stream in modal.
The modal that appears on entering a room creates a new MediaStream
object for the video preview every time a video device is selected. Now
we stop the associated tracks of the MediaStream when another device is
selected and when the component unmounts. This allows the camera to turn
off and prevents having unused tracks running.