Fix streaming bugs.
Occasionally audio or video would not show up on the other end of the
connection. There were two issues causing this. First, a track might
come in that has no associated stream, so the stream array that comes
with the track is actually empty. We now check for streams[0] before
updating feed.stream, which prevents the <video> srcObject from being
set to undefined. Second, it seems that <video> will refuse to play an
audio track if it is also given a muted video track. This is remedied by
having separate <video> and <audio> elements for each VideoFeed so that
audio will play regardless of what is happening with video.

In a small refactoring, starting and stopping the manager has been moved
from chat.js to App.js.