Remove stream on remote track end.
Previously when a remote video track was stopped, there would remain a
frozen image in the video feed. We now detect when a remote track has
ended by listening for the mute event, then remove the stream from the
corresponding <video> or <audio> element. This prevents the frozen image
and instead displays black when a remote video track has been stopped.