Introduce Redux to manage React state.
Redux allows us to centralize global state. This should turn out better
than having global state stored in the top-level App component, since
each component can access the state that it needs to directly rather
than passing props and callback functions through the component
hierarchy.

Updating the manager (e.g. setting a username) is considered a
side-effect, so Redux-Saga will be used as the point of contact between
the manager and the rest of the front end.