Add video resolution controls.
This commit adds resolution controls for both outgoing and incoming
video. However, it turns out that resolution controls for incoming video
are mostly unnecessary since WebRTC automatically downscales resolution
based on resource constraints. The implementation in this commit
fundamentally works in allowing users to downscale resolution, but makes
the mistake of using the original track resolution (rather than the
automatically downscaled resolution) to manually compute a downscaling
factor. This results in downscaling video on the sender side twice, once
by us and once by the WebRTC encoder.

This commit is basically a dead-end, but contains some useful code for
later reference. While the resolution controls for incoming video will
be scrapped, resolution controls for outgoing video are still useful.