Implement adding ICE servers in UI.
An `enabled` property has been added to ICE server objects in the Redux
store. This allows us to add empty servers in the UI without passing
them to the manager. Then, when a server is saved, it is enabled and can
be passed to the manager. Validation (i.e. checking the format of urls
and requiring username/credential for TURN servers) will be added in a
later commit.