Add validation for ICE server forms.
We use basic HTML form validation with the `pattern` and `required`
attributes on form inputs. The URLs input must be a comma-separated list
of URLs, each prefixed with the server type (i.e. stun: or turn:).
Username and password are required fields for TURN servers, but not
required for STUN servers.