Enable video on room entry if device was selected.
Bugfix. Devices selected in EnterRoomModal should be enabled immediately
after the modal is closed. A recent commit broke this behavior for
video. Now when a camera is selected in the modal, it is enabled in the
room.

Refactor -- MediaControlBar now uses the `active` property of devices
objects in Redux to determine whether devices are on or off, instead of
tracking this with local state.