Fix unnecessary scrollbar issue.
On the initial render of the VideoStage, the height reported by the
`clientHeight` attribute is slightly larger than it should be, for
unknown reasons (the `clientHeight` is about 8px less on subsequent
renders). For the first person entering the room, this sometimes results
in y-overflow for the VideoStage and causes a scrollbar to be displayed.
The simplest solution for now is to change the slack constant from 2px
to 4px in calculateScaleFactor().