Change client reaping time from 60 to 30 seconds.
Websockets tend to time out after about 60 seconds depending on the
webserver. On Heroku, the timeout is not configurable. This was causing
clients to be reaped because the websocket was being closed, preventing
the ping/pong exchange. Since we now ping the client every 30 seconds,
the websocket shouldn't time out.